testing will be done with asserts
Since vs2012 is not compiler independent it can't be used. Furthermore
cppunit is much too complicated for our uses plus needs to be build as a
lib which cant be easily done automatically it can not be used. The
feasible alternative is to just use BOOST_ASSERT and create a simple
executable for testing.
This is done for the first testclass.